+++ clease/montecarlo/base.py
@@ -1,6 +1,7 @@
 from typing import Union
 from abc import ABC
 from ase import Atoms
+from ase.units import kB
 from .mc_evaluator import MCEvaluator, construct_evaluator
 
 __all__ = ("BaseMC",)
@@ -57,6 +58,7 @@ class BaseMC(ABC):
     def _set_temperature(self, value: float) -> None:
         """Set the internal temperature."""
         self._temperature = value  # pylint: disable=attribute-defined-outside-init
+        self.kT = value * kB  # pylint: disable=attribute-defined-outside-init
         self._on_temp_change()
 
     def _on_temp_change(self) -> None: